PostOffice/PostOffice/requirements.txt
Django==1.7.5
djangorestframework==3.0.5
ecdsa==0.13
Fabric==1.10.1
paramiko==1.15.2
Paver==1.2.4
django-allauth==0.19.1
django-rest-auth==0.3.4
